requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cobble==0.1.4
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
mammoth==1.8.0
markdownify==0.14.1
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
pandas==2.2.3
pathvalidate==3.2.1
pdfminer.six==20240706
pillow==11.0.0
puremagic==1.28
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.12.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
XlsxWriter==3.2.0
youtube-transcript-api==0.6.3
